fix: endpoint 하나로 만들고 필요없는 파일 삭제
